util.sh: sudoku: cache de sdk_ecris
+ On essaie de cacher le résultat de dernière exécution, pour économiser des écritures / lectures / suppressions.

darcs-hash:b72282356f709d9ddf668f1d3fd6c2e54b733a1b

+++ util.sh
@@ -437,7 +437,11 @@ sudoku()
 		# Qu'a-t-on à notre disposition?
 		
 		# sdk_ecris: peut-on écrire dans le dossier cible?
+		if [ -z "$sdk_ecris" -o -z "$sdk_ecrisDernierOu" -o "x$sdk_ecrisDernierOu" != "x$ou" ] # Un minimum de cache.
+		then
 		touch "$ou/.sudoku.temoin" 2> /dev/null && rm -f "$ou/.sudoku.temoin" 2> /dev/null && sdk_ecris=1 || sdk_ecris=0
+			sdk_ecrisDernierOu="$ou"
+		fi
 		# sdk_vraiment: veut-on *vraiment* passer root?
 		# Du fait de notre malheureuse surcharge de sudo, parfois notre sudoku est appelé pour faire du vrai sudo où on a vraiment besoin de passer root (même si $INSTALLS est inscriptible; par exemple pour ajouter à /usr/local/etc/rc.d un lien symbolique vers un de nos fichiers "à nous"). Pour cette usage, appeler SANSSU=0 sudoku ….
 		[ "x$SANSSU" = x0 ] && sdk_vraiment=1 || sdk_vraiment=0
